move card component into individual files
