+++ src/components/CohortBuilder/Summary/AgeDiagChart.js
@@ -4,20 +4,23 @@ import { withTheme } from 'emotion-theming';
 import { get, sumBy } from 'lodash';
 import gql from 'graphql-tag';
 import VerticalBar from 'chartkit/components/VerticalBar';
+import { CohortCard } from './ui';
 
 const ageAtDiagnosisTooltip = data => {
   return `${data.value.toLocaleString()} Participant${data.value > 1 ? 's' : ''}`;
 };
 
 const AgeDiagChart = ({ data, theme }) => (
-  <VerticalBar
-    data={data}
-    indexBy="label"
-    tooltipFormatter={ageAtDiagnosisTooltip}
-    sortByValue={true}
-    height={225}
-    colors={[theme.chartColors.lightblue]}
-  />
+  <CohortCard title="Age at Diagnosis">
+    <VerticalBar
+      data={data}
+      indexBy="label"
+      tooltipFormatter={ageAtDiagnosisTooltip}
+      sortByValue={true}
+      height={225}
+      colors={[theme.chartColors.lightblue]}
+    />
+  </CohortCard>
 );
 
 export const ageDiagQuery = sqon => ({
