+++ src/components/CohortBuilder/Summary/DemographicChart.js
@@ -5,6 +5,7 @@ import { withTheme } from 'emotion-theming';
 import { compose } from 'recompose';
 import { get, camelCase } from 'lodash';
 import Pie from 'chartkit/components/Pie';
+import { CohortCard } from './ui';
 
 const PieChartContainer = styled('div')`
   display: flex;
@@ -19,33 +20,35 @@ const PieChartContainer = styled('div')`
 `;
 
 const DemographicChart = ({ data, theme }) => (
-  <PieChartContainer>
-    <Pie
-      style={{ height: '42%', width: '50%', marginBottom: '10px', marginTop: '5px' }}
-      title={'Gender'}
-      data={data.gender}
-      colors={[theme.chartColors.orange, '#FFFFFF']}
-    />
-    <Pie
-      style={{ height: '42%', width: '50%', marginBottom: '10px', marginTop: '5px' }}
-      title={'Ethnicity'}
-      data={data.ethnicity}
-      colors={[theme.chartColors.darkblue, '#FFFFFF']}
-    />
-    <Pie
-      style={{ height: '42%', width: '50%' }}
-      title={'Race'}
-      data={data.race}
-      colors={[theme.chartColors.lightpurple, '#FFFFFF']}
-    />
+  <CohortCard showHeader={false}>
+    <PieChartContainer>
+      <Pie
+        style={{ height: '42%', width: '50%', marginBottom: '10px', marginTop: '5px' }}
+        title={'Gender'}
+        data={data.gender}
+        colors={[theme.chartColors.orange, '#FFFFFF']}
+      />
+      <Pie
+        style={{ height: '42%', width: '50%', marginBottom: '10px', marginTop: '5px' }}
+        title={'Ethnicity'}
+        data={data.ethnicity}
+        colors={[theme.chartColors.darkblue, '#FFFFFF']}
+      />
+      <Pie
+        style={{ height: '42%', width: '50%' }}
+        title={'Race'}
+        data={data.race}
+        colors={[theme.chartColors.lightpurple, '#FFFFFF']}
+      />
 
-    <Pie
-      style={{ height: '42%', width: '50%' }}
-      title={'Family Composition'}
-      data={data.familyComposition}
-      colors={[theme.chartColors.lightblue, '#FFFFFF']}
-    />
-  </PieChartContainer>
+      <Pie
+        style={{ height: '42%', width: '50%' }}
+        title={'Family Composition'}
+        data={data.familyComposition}
+        colors={[theme.chartColors.lightblue, '#FFFFFF']}
+      />
+    </PieChartContainer>
+  </CohortCard>
 );
 
 export const demographicQuery = sqon => ({
